fixed warnings

+++ src/gateway/preBuffer/Dispatcher.cc
@@ -100,7 +100,7 @@ void Dispatcher::handleMessage(cMessage *msg)
         if(strcmp(dispatchedBackboneTransferType.c_str(), "BE") == 0){
             interDataStructure->setDirectMacAdress((sendBuffer->getDispatchedCTID()).c_str());
         }else{
-            interDataStructure->setBackboneCTID(atoi((sendBuffer->getDispatchedCTID()).c_str()));
+            interDataStructure->setBackboneCTID( ((uint16_t)atoi((sendBuffer->getDispatchedCTID()).c_str())) );
         }
         interDataStructure->encapsulate(multiFieldSequence);
         send(interDataStructure, "moduleConnect$o");
